Re-implement legacy serialization
